src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/exposed-object-within-shadow-expected.txt

CONSOLE ERROR: line 19: Uncaught TypeError: host.createShadowRoot is not a function
Shadow DOM inside an <object> and a <div>
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
